refactor: use class instead of function
