+++ package.json
@@ -16,7 +16,8 @@
     "local:trigger-lint-on-commit": "husky install",
     "local:prevent-trigger-lint-on-commit": "git config --unset core.hooksPath",
     "preinstall": "npx check-engine",
-    "test": "jest"
+    "test": "jest",
+    "start": "nodemon src/lib/index.ts"
   },
   "repository": {
     "type": "git",
@@ -50,6 +51,7 @@
     "husky": "^8.0.3",
     "jest": "^29.5.0",
     "lint-staged": "^13.2.2",
+    "nodemon": "^2.0.22",
     "prettier": "^3.0.1",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
